+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.2
+* Fix encode/decode for non-empty Era/AnonymousStruct
+* Add multiple convertion for Extrinsics
+* Add encode/decode for IdentityInfoAdditional/IdentityInfo/Null/Data
+* Add encode/decode for MetadataV11
+* Add README for examples
+* Add coverage test
+* Add test cases for Extrinsics/Era/AnonymousStructure
+
 # 0.1.1
 * Fix special encode/decode rule for Option\<Bool\>.
 * Fix encode/decode for signed integers to avoid overflow.